Store user token instead of username
Previously, the frontend remembered the active user by storing their
username in localStorage.  This means that a savvy user could
impersonate another user by simply changing the value in their browser.

To improve security, this has been replaced with an account token
generated by the backend server.  Because it depends on a secret stored
in the backend, it cannot be forged; the only way to obtain a valid
token is to login legitimately.

To obtain data about the user, the frontend requests /userdata from the
backend, which returns all the data as JSON.